OpCode
Completed opcode 90 percent
included random byte generator as there is an instruction which places a random number into a register.
Some opcodes are still not completed as they includes display and keyboard instructions so I thought when display and keyboard will gets completed then I will add the code for it.
some declarations are changed as austins method of declaration seems nice and professional so i copied and learned next time i will use this method for declaration.
I also made different file for opcode table for its decoding and coding part .

Please check it if  I have declared all the variables right and connected all the files in right way or not.